ds - moved EqSearches.java into searches directory
